backend/requirements.prod.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.2.1
pybit==5.14.0
pydantic==2.12.5
numpy==2.4.2
pandas==3.0.1
lightgbm==4.6.0
scikit-learn==1.8.0
joblib==1.5.3
scipy==1.17.1
python-multipart==0.0.22
aiohttp==3.13.3
websockets==16.0
cryptography==46.0.5
httpx==0.28.1
starlette==0.37.2
anyio==4.12.1
